cheri: usb: class: Fix ioctls in drivers/usb/class for CHERI
Fix CHERI warnings and the ioctl type for drivers/usb/class/.

+++ drivers/usb/class/cdc-wdm.c
@@ -788,7 +788,7 @@ static int wdm_release(struct inode *inode, struct file *file)
 	return 0;
 }
 
-static long wdm_ioctl(struct file *file, unsigned int cmd, unsigned long arg)
+static long wdm_ioctl(struct file *file, unsigned int cmd, uintptr_t arg)
 {
 	struct wdm_device *desc = file->private_data;
 	int rv = 0;
